# user-management-service/src/main/java/com/ps/userservice/dto/LoginResponse.java
package com.ps.userservice.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class LoginResponse {
    private String token;
}